backend/requirements.txt
fastapi
uvicorn[standard]
msal
requests
jinja2
python-dotenv
python-multipart